refactor: split URL service into management and redirect services, extract validator
