--- src/main/java/com/rapidlink/services/ShortUrlService.java
@@ -1,10 +1,0 @@
-package com.rapidlink.services;
-
-import java.net.URI;
-
-public interface ShortUrlService {
-
-    String createShortUrl(String originalUrl);
-
-    URI resolveUrl(String shortCode);
-}
