+++ src/main/java/com/rapidlink/services/UrlManagementService.java
@@ -1,0 +1,5 @@
+package com.rapidlink.services;
+
+public interface UrlManagementService {
+    String createUrl(String originalUrl);
+}
