+++ src/main/java/com/rapidlink/services/UrlRedirectService.java
@@ -1,0 +1,7 @@
+package com.rapidlink.services;
+
+import java.net.URI;
+
+public interface UrlRedirectService {
+    URI getRedirectUrl(String shortCode);
+}
